+++ openbas-sentinel/src/main/java/io/openbas/collectors/sentinel/domain/Alert.java
@@ -1,0 +1,13 @@
+package io.openbas.collectors.sentinel.domain;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Data;
+
+@Data
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class Alert {
+
+  private String kind; // SecurityAlert for us
+  private AlertProperties properties;
+
+}
