+++ openbas-sentinel/src/main/java/io/openbas/collectors/sentinel/domain/ListOfEntities.java
@@ -1,0 +1,14 @@
+package io.openbas.collectors.sentinel.domain;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.Data;
+
+import java.util.List;
+
+@Data
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class ListOfEntities {
+
+  private List<Entity> entities;
+
+}
