+++ openbas-sentinel/src/main/java/io/openbas/collectors/sentinel/utils/Utils.java
@@ -1,0 +1,30 @@
+package io.openbas.collectors.sentinel.utils;
+
+import java.time.Instant;
+import java.time.LocalDateTime;
+import java.time.ZonedDateTime;
+import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeFormatterBuilder;
+import java.time.temporal.ChronoField;
+import java.util.Locale;
+
+import static java.time.ZoneOffset.UTC;
+
+public class Utils {
+
+  public static final DateTimeFormatter FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss'z'", Locale.FRANCE);
+
+  public static Instant toInstant(final String date) {
+    DateTimeFormatter dateTimeFormatter = new DateTimeFormatterBuilder()
+        .appendPattern("yyyy-MM-dd'T'HH:mm:ss")
+        .optionalStart()
+        .appendFraction(ChronoField.NANO_OF_SECOND, 0, 9, true)
+        .optionalEnd()
+        .appendPattern("'Z'")
+        .toFormatter();
+    LocalDateTime localDateTime = LocalDateTime.parse(date, dateTimeFormatter);
+    ZonedDateTime zonedDateTime = localDateTime.atZone(UTC);
+    return zonedDateTime.toInstant();
+  }
+
+}
