+++ openbas-sentinel/src/test/java/io/openbas/collectors/sentinel/App.java
@@ -1,0 +1,15 @@
+package io.openbas.collectors.sentinel;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+@EnableScheduling
+@SpringBootApplication
+public class App {
+
+  public static void main(String[] args) {
+    SpringApplication.run(App.class, args);
+  }
+
+}
