+++ openbas-sentinel/src/test/resources/application.properties
@@ -1,0 +1,31 @@
+# OpenBAS configuration
+openbas.base-url=http://localhost:8080
+
+# Database Properties
+spring.datasource.url=jdbc:postgresql://localhost:5432/postgres
+spring.datasource.username=postgres
+spring.datasource.password=admin
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+
+# Minio Properties
+minio.endpoint=localhost
+minio.port=10000
+minio.bucket=openbas
+minio.access-key=minioadmin
+minio.access-secret=minioadmin
+
+## Collector Sentinel
+collector.sentinel.enable=true
+collector.sentinel.id=04d26c87-16a8-46b4-b3bb-1ee5a48a0389
+collector.sentinel.authority.base-url: https://login.microsoftonline.com
+collector.sentinel.authority.tenant-id: f2a29c7e-5230-41d3-b5d1-4b6e2c4e6b7c
+collector.sentinel.authority.url: ${collector.sentinel.authority.base-url}/${collector.sentinel.authority.tenant-id}
+collector.sentinel.client-id: 726e781b-6215-4119-a4cd-9d43ef16c606
+collector.sentinel.secret: MvJ8Q~AJMPQpkQ.q5H9YXsXfF~HNmKJfNxHe3be2
+collector.sentinel.scope: https://management.azure.com/.default
+collector.sentinel.endpoint.workspace.subscription-id: 4588a669-597f-4097-b51f-274b496efa4d
+collector.sentinel.endpoint.workspace.resourcegroup-name: default
+collector.sentinel.endpoint.workspace.name: Filigran
+collector.sentinel.endpoint.workspace.base-url: subscriptions/${collector.sentinel.endpoint.workspace.subscription-id}/resourcegroups/${collector.sentinel.endpoint.workspace.resourcegroup-name}/providers/microsoft.operationalinsights/workspaces/${collector.sentinel.endpoint.workspace.name}
+collector.sentinel.endpoint.api-version: 2023-12-01-preview
+collector.sentinel.endpoint.url: https://management.azure.com/${collector.sentinel.endpoint.workspace.base-url}/providers/Microsoft.SecurityInsights/
