--- openex-sentinel/src/main/java/io/openex/collectors/sentinel/application/SentinelJob.java
@@ -1,17 +1,0 @@
-package io.openex.collectors.sentinel.application;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@RequiredArgsConstructor
-@Service
-public class SentinelJob implements Runnable {
-
-  private final SentinelService sentinelService;
-
-  @Override
-  public void run() {
-    this.sentinelService.computeDetectionExpectations();
-  }
-
-}
