--- openex-sentinel/src/main/java/io/openex/collectors/sentinel/domain/Alert.java
@@ -1,13 +1,0 @@
-package io.openex.collectors.sentinel.domain;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.Data;
-
-@Data
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Alert {
-
-  private String kind; // SecurityAlert for us
-  private AlertProperties properties;
-
-}
