--- openex-sentinel/src/main/java/io/openex/collectors/sentinel/domain/Entity.java
@@ -1,13 +1,0 @@
-package io.openex.collectors.sentinel.domain;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.Data;
-
-@Data
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class Entity {
-
-  private String kind; // Host for us
-  private HostProperties properties;
-
-}
