--- openex-sentinel/src/main/java/io/openex/collectors/sentinel/domain/ListOfAlerts.java
@@ -1,14 +1,0 @@
-package io.openex.collectors.sentinel.domain;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import lombok.Data;
-
-import java.util.List;
-
-@Data
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class ListOfAlerts {
-
-  private List<Alert> value;
-
-}
