--- openex-sentinel/src/main/java/io/openex/collectors/sentinel/infrastructure/WebConfig.java
@@ -1,18 +1,0 @@
-package io.openex.collectors.sentinel.infrastructure;
-
-import org.springframework.boot.web.client.RestTemplateBuilder;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.client.RestTemplate;
-
-import java.time.Duration;
-
-@Configuration
-public class WebConfig {
-  @Bean
-  public RestTemplate restTemplate(RestTemplateBuilder builder) {
-    return builder
-        .setConnectTimeout(Duration.ofMillis(3000))
-        .build();
-  }
-}
