--- openex-sentinel/src/test/java/io/openex/collectors/sentinel/App.java
@@ -1,15 +1,0 @@
-package io.openex.collectors.sentinel;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.scheduling.annotation.EnableScheduling;
-
-@EnableScheduling
-@SpringBootApplication
-public class App {
-
-  public static void main(String[] args) {
-    SpringApplication.run(App.class, args);
-  }
-
-}
