--- openex-sentinel/src/test/java/io/openex/collectors/sentinel/application/SentinelServiceTest.java
@@ -1,18 +1,0 @@
-package io.openex.collectors.sentinel.application;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-public class SentinelServiceTest {
-
-  @Autowired
-  private SentinelService sentinelService;
-
-  @Test
-  void test() {
-    this.sentinelService.computeDetectionExpectations();
-  }
-
-}
